[TASK] Object Factory first implementation
Summary:
Initial implementation of an extension-native
ObjectFactory to load objects. This first implementation
uses Extbase's ObjectManager to load objects and returns
a dummy MasterConfiguration which will only return
default values.

Intention is to allow this ObjectFactory to create other
class names based on registered implementations without
depending on a TYPO3CMS or Flow3 specific method of
doing so. It also allows custom initialization of some
classes again without depending on a framework-specific
strategy - and makes a central point for adapting the
class loading, by letting the framework load another
custom-purpose ObjectFactory.

While it is not the main goal of this pattern, it also
helps unit test writing in that the factory can be mocked
to return other mocks with proper expectations.

Reviewers: dkd-egerer, #forgetit-dev

Reviewed By: dkd-egerer, #forgetit-dev

Differential Revision: https://phabricator.dkd.de/D153