[FEATURE] Implement an initial SimpleQueue
Summary:
WORD OF WARNING:

I am aware that this current implementation has some obvious technical
flaws, mainly in that it uses a cache table to store the entire queue
as one big object, inevitably causing problems on large queues and
due to the nature of caches, gets cleared when "system" caches are
cleared.

The final SimpleQueue should use plain records in a custom table, but
this first implementation allows work to continue until the other parts
are better established and the queue implementation can be optimised.

As such, it is WIP but I would still like to merge it at this point
with the promise of a more solid implementation coming shortly after
completing the CommandController / Scheduler parts that actually use
this Queue.

I'm of course open to discuss the final implementation of SimpleQueue
that I plan to do based on a table without TCA. Also a possible topic
would be whether or not we should implement one or two other Queue
types - or just stick with a DB-based single implementation.

Reviewers: dkd-egerer, #forgetit-dev

Reviewed By: dkd-egerer, #forgetit-dev

Differential Revision: https://phabricator.dkd.de/D200