[FEATURE] Execution-context validation of Tasks
Summary:
This change adds a validate(TaskInterface) method to Execution
allowing each type of Execution to validate the Task they are
given. If an Execution knows in advance it will not be able to
correctly execute the Task, an InvalidArgumentException can be
raised from within the validate() method and the Worker (in
our current default implementation) will catch this and create
a friendlier Result that contains the error and its message.
The returned Result has the code Result::ERR (int) 1 and the
original error as payload.

Reviewers: dkd-egerer, #forgetit-dev

Reviewed By: dkd-egerer, #forgetit-dev

Differential Revision: https://phabricator.dkd.de/D296