[FEATURE] Option on indexing task to allow double pass indexing
Summary:
Adds a "relations" option on RecordIndexTask which when FALSE
(which is default) will NOT index relations from one object
to another. This allows indexing to happen in two passes:
First pass to ensure all tracked objects are indexed.
Second pass to fill in UUID of related records indexed in step 1.

Reviewers: #forgetit-dev, dkd-egerer

Reviewed By: #forgetit-dev, dkd-egerer

Differential Revision: https://phabricator.dkd.de/D324