[TASK] Let commands use "cmis:" prefix and add a dump command
Summary:
This renames the CommandController from using the "cronjob:"
prefix to using a more appropriate "cmis:" prefix.

Also added is an additional command for the controller which
dumps selected resources; configuration or tree currently
can be dumped, with tree supporting full and brief modes.

Test Plan:
arc unit
Currently no unit tests cover this code since the treatment
of the Tree is expected to change later.

Reviewers: dkd-ebert, #forgetit-dev

Reviewed By: dkd-ebert, #forgetit-dev

Differential Revision: https://phabricator.dkd.de/D1006