[TASK] Implement a database-based queue as default
Summary:
This queue will perform much, much better than the
current (testing-grade) queue implementation. The new
Queue type is defined as default.

Task is extended with a getResourceId() method which
can return a unique identifier for resources related
to Tasks, allowing SQL queries to quickly check if
there already is an identical Task for that resource,
thus preventing duplicates.

Methods relevant only to "file-cachable" queues have
been moved to the QueueCachableInterface instead.

Johnny added as reviewer to make sure the SQL schema
is added to the docker images if needed!

Reviewers: dkd-goslar, #[t]_forgetit-dev, dkd-ebert

Reviewed By: #[t]_forgetit-dev, dkd-ebert

Differential Revision: https://phabricator.dkd.de/D1646