[TASK] Register "filemetadata" extension as dependency
Summary:
The extension adds the fields we touch when automatically
indexing FAL (and relations to FAL objects). Cannot be in
`cmis_fal` because this extension is optional and the
fields are touched even when the "filemetadata" extension
is not installed.

The touching of fields even when "filemetadata" is not
installed may very well be a TYPO3 core bug - but, we can
very easily work around it with this simple added dep.

Reviewers: dkd-ebert, dkd-goslar, #[t]_forgetit-dev

Reviewed By: dkd-goslar, #[t]_forgetit-dev

Differential Revision: https://phabricator.dkd.de/D1736