[TASK] Create a hostname-based Site for global records
Summary:
When records exist in page UID zero (globally) they are
impossible to correctly associate with a TYPO3 domain,
instead, the hostname of the current host is resolved
and used as folder. We use the `hostname` command because
we don't necessarily have an HTTP request and would like
to avoid putting global records into two different folders
when the indexing is called via CLI and HTTP respectively.

Reviewers: dkd-goslar, #[t]_forgetit-dev, dkd-ebert

Reviewed By: #[t]_forgetit-dev, dkd-ebert

Differential Revision: https://phabricator.dkd.de/D1771