[TASK] Configure sys_file indexing for non-CMIS files
Summary:
This adds the basic ability to index sys_file records pointing to
files which are NOT currently stored in CMIS and any relations to
such files.

The CMIS FAL extension takes care of indexing relations to CMIS
files which are consumed through a FAL CMIS mount point.

Properties of sys_file are not yet mapped; the current state is
intended to allow "dumb" relations like "record references file".

Reviewers: dkd-goslar

Reviewed By: dkd-goslar

Differential Revision: https://phabricator.dkd.de/D2076