[FEATURE] Record import based on reverse index mapping
This change adds an importing task which can import identified CMIS objects to TYPO3 records with the target table and column values being decided by the indexing configuration, but reversed.

Also adds version tracking to TYPO3's cache of identities to store the currently known version of the CMIS object. The version number in cache can then be used in coming changes to perform merging of values or preferring the CMIS or TYPO3 version of data.

Triggered by detecting new CMIS objects recursively using the `cmis:generateimportingtasks` command - which queues tasks that are executed with the usual `cmis:picktask` or `cmis:picktasks` or backend module.

Note: can be tested by indexing all pages and news items, then deleting a news item and also deleting it from the recycler - then running the import task. Since the CMIS object now has no corresponding TYPO3 record, an import task is generated and when executed, the news record is restored with all indexed values. Alternatively, manually create a CMIS object of a type that is indexed and let the generateimportingtasks command find the new object.