[BUGFIX] Add new server-configuration cache mechanism
The server credentials where cached in a yaml file. But this file
is removed when the cache is cleared.
When using CMIS in the frontend it is sometimes used in eid scripts.
Eid scripts do not have TypoScript loaded so the credentials can't
be read.
Therefore the credentials are now put into the TYPO3 registry by
clicking the "Initialize CMIS server" button in the cache menu.
This implementation currently lacks support for mulit-page-tree
implementations but that was also the case before because there
was only one cache file for all page trees. So it doesn't matter
in which tree the cmis server has been configured but it does matter
what the cache file contains. Now it does matter what the registry
contains.